[rend2] Fix shader remaps
Theres some optimizations for the prepass to merge more surfaces which works fine for opaque shaders and remaps that are also opaque. When changing to transparent shaders those optimization didnt work properly as they didnt know about the remap. Remapping now happens when drawsurfaces are added and not when drawing of the surfaces starts.